requirements.txt
requests
beautifulsoup4
pandas
python-dotenv
supabase
flask
flask-cors
google-genai
